update kernel
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@59678 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-4.4/fs/nilfs2/dir.c
@@ -67,7 +67,7 @@ static inline void nilfs_put_page(struct page *page)
  */
 static unsigned nilfs_last_byte(struct inode *inode, unsigned long page_nr)
 {
-	unsigned last_byte = inode->i_size;
+	u64 last_byte = inode->i_size;
 
 	last_byte -= page_nr << PAGE_CACHE_SHIFT;
 	if (last_byte > PAGE_CACHE_SIZE)
